+++ iterm/setup_iterm.sh
@@ -1,9 +1,11 @@
 
 function setup_iterm(){
-  if ! [[ -e /Applications/iTerm.app ]]; then
+  if ! [[ -e $HOME/Applications/iTerm.app ]] && ! [[ -e /Applications/iTerm.app ]]; then
     print_with_color $GREEN 'iTerm is not installed. Downloading...'
     brew install Caskroom/cask/iterm2
   fi
+
+  check_and_link_file `pwd`/iterm/com.googlecode.iterm2.plist /Users/mat/Library/Preferences/com.googlecode.iterm2.plist
 }
 
 print_with_color $YELLOW 'Do you want to install iTerm? (yes/no)'
